requirements.txt
Flask
Flask-Login
Werkzeug
tensorflow
Pillow
python-dotenv
psycopg2-binary
cloudinary
gunicorn
requests